+++ app/src/main/java/com/work/terry/snakeonastring_jbox2d/SurfaceViewAndActivity/StartView.java
@@ -703,7 +703,7 @@ public class StartView extends MyView {
     }
 
     @Override
-    public void onTouchEvent(MotionEvent event, int x, int y) {
+    public void onTouchEvent(MotionEvent event, float x, float y) {
         switch (event.getAction()){
             case MotionEvent.ACTION_DOWN:
             case MotionEvent.ACTION_MOVE:
